pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "tg-dog"
version = "0.1.0"
description = "Telegram userbot CLI built with Telethon and APScheduler"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "sunnysab <i@sunnysab.cn>" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "typer>=0.12.0",
  "telethon>=1.34.0",
  "pyyaml>=6.0.1",
  "apscheduler>=3.10.4",
  "tqdm>=4.66.0",
  "PySocks>=1.7.1",
]

[project.scripts]
tg-dog = "main:app"

[tool.hatch.build.targets.wheel]
packages = ["core"]
include = ["main.py", "plugins/**", "README.md", "config.yaml"]
